+++ packages/components/CHANGELOG.md
@@ -20,6 +20,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
    ([#4285](https://github.com/porsche-design-system/porsche-design-system/pull/4285))
 - `Accessibility`: improved internal ARIA attribute handling for various components, e.g. `Banner`, `Inline Notification` and `Input` elements.
   ([#4274](https://github.com/porsche-design-system/porsche-design-system/pull/4274))
+- Prefixing: custom PDS prefixes allow numbers in the prefix name
+  ([#4297](https://github.com/porsche-design-system/porsche-design-system/pull/4297))
 
 ## [4.0.0-rc.1] - 2026-04-07