+++ packages/components/CHANGELOG.md
@@ -22,6 +22,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
    ([#4285](https://github.com/porsche-design-system/porsche-design-system/pull/4285))
 - `Accessibility`: improved internal ARIA attribute handling for various components, e.g. `Banner`, `Inline Notification` and `Input` elements.
   ([#4274](https://github.com/porsche-design-system/porsche-design-system/pull/4274))
+- `Tailwind CSS`: Custom tailwind prefixing
+  ([#4294](https://github.com/porsche-design-system/porsche-design-system/pull/4294))
 
 ## [4.0.0-rc.1] - 2026-04-07