+++ packages/components/CHANGELOG.md
@@ -26,7 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
   ([#4297](https://github.com/porsche-design-system/porsche-design-system/pull/4297))
 - `Tailwind CSS`: Custom tailwind prefixing
   ([#4294](https://github.com/porsche-design-system/porsche-design-system/pull/4294))
-- Form components: rendering issues of slotted label if used as `display: block` elements 
+- Form components: rendering issues of slotted label if provided as `display: block` element
   ([#4298](https://github.com/porsche-design-system/porsche-design-system/pull/4298))
 
 ## [4.0.0-rc.1] - 2026-04-07