+++ packages/components/CHANGELOG.md
@@ -14,7 +14,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ## [Unreleased]
 
-- `OptGroup`, `Multi Select`, `Select`: added internal event to detect slots in `OptGroup`
+### Fixed
+
+- `Multi Select`, `Select`: not correctly updating when options within `p-optgroup` change dynamically
   ([#4278](https://github.com/porsche-design-system/porsche-design-system/pull/4279))
 
 ## [3.34.0] - 2026-04-09