+++ packages/components/CHANGELOG.md
@@ -36,6 +36,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
   ([#4298](https://github.com/porsche-design-system/porsche-design-system/pull/4298))
 - `Tabs`, `Tabs Bar`: keyboard navigation in RTL mode
   ([#4300](https://github.com/porsche-design-system/porsche-design-system/pull/4300))
+- `Select`, `Multi Select`: Optimize screen reader experience
+  ([#4313](https://github.com/porsche-design-system/porsche-design-system/pull/4313))
 
 ## [4.0.0-rc.1] - 2026-04-07