+++ packages/components/CHANGELOG.md
@@ -37,6 +37,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - `Tabs`, `Tabs Bar`:
   - keyboard navigation in RTL mode ([#4300](https://github.com/porsche-design-system/porsche-design-system/pull/4300))
   - animation flickering ([#4309](https://github.com/porsche-design-system/porsche-design-system/pull/4309))
+  - hydration error ([#4309](https://github.com/porsche-design-system/porsche-design-system/pull/4309))
 
 ## [4.0.0-rc.1] - 2026-04-07