+++ packages/components/CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ## [Unreleased]
 
+## [4.0.0-rc.2] - 2026-04-28
+
 ### Added
 
 - `Tag`: Prop `variants` supports new values: